requirements.txt
transformers==4.40.2
torch==2.30
pillow==10.3.0
fastapi==0.111.0